【修改】修改专家品目的不通过理由回显【作者】王慧杰

+++ src/main/java/ses/service/ems/impl/ExpertServiceImpl.java
@@ -672,7 +672,7 @@ public class ExpertServiceImpl implements ExpertService {
       if(auditList!=null && auditList.size()>0){
         for (ExpertAudit expertAudit : auditList) {
           //修改之前的审核信息为删除 和历史状态
-          expertAudit.setIsDelete((long) 1);
+          expertAudit.setIsDelete(1);
           expertAudit.setIsHistory("1");
           expertAuditMapper.updateByPrimaryKeySelective(expertAudit);
         }
